Refactoriza estructura de la aplicaion
