+++ src/components/layout/AppLayout.tsx
@@ -125,7 +125,7 @@ const TabAddAction = (): JSX.Element => {
 
 const AppLayout = ({ children }: AppLayoutProps): JSX.Element => {
   return (
-    <div className="mx-auto flex h-screen w-full flex-col sm:max-w-lg">
+    <div className="mx-auto flex h-full w-full flex-col sm:max-w-lg">
       <main className="grow overflow-y-scroll">{children}</main>
       <TabBarMenu />
     </div>
